test: should update state and id when a new item is received from input box
